zzw2

+++ src/main/java/org/example/smarthospital_back/entity/UserEntity.java
@@ -1,7 +1,7 @@
 package org.example.smarthospital_back.entity;
 
 public class UserEntity {
-    long u_id;
+    int u_id;
     long u_tel;
     int u_age;
     String u_id_card;
